feature-init-routing
+ Split App to page components
