+++ src/pages/About.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const About = () => {
+	return (
+		<div>
+			<h1 style={{
+				textAlign: 'center'
+			}}> About page </h1>
+		</div>
+	)
+}
+
+export default About
